aarch64: make queue_mpsc work with weak memory model
This patch changes the queue_mpsc to make the poplist variable an atomic.
The poplist is intended to be manipulated by single thread at any point
in time (single consumer) which means that do not need to protect it from regular races.
However the poplist consumers (even when one at the time) may run on different CPUs
and therefore need to SEE the latest value of this variable
after the write on a different CPUs. For that reason we make poplist
an atomic and use acquire/release semantics to archieve cross-CPU visibility.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>