aarch64: make java-base and related modules compile
This patch does more than what the title suggests and ideally
should be split in at least 2 parts. But it would require substantial
amount of work to untangle those and more importantly re-test at this point.

So firstly, this patch modifies makefiles and module.py files
of java related modules and others like golang that include
modules/java-base/common.gmk to make it possible to build
on aarch64 host. It achieves it mostly by including modules/common.gmk
which provides necessary rules for compiling on both x64 and aarch64 hosts
as well as cross-compiling aarch64 on x64 host. This also removes
a lot of repeated boiler plate and makes many changed makefiles
more consistent with each other. Finally all build artifacts
like object files are output to
./build/<release|debug>.[arch]/modules/[module] directories instead of
modules/[module].

Secondly, this patch also modifies the build process to dynamically
determine which jdk to use and which java tests to run. Therefore
new version of modules/java/module.py detects which version of java
(8 or 9 and above) is in the PATH and accordingly selects openjdk8-from-host
or openjdk9_1x-from-host. In addition, we also make java-tests
makefile generate list of java tests to be executed when test.py is run.
This is necessary because our java "wrapper" mechanism is no longer
compatible with Java 9 and above and we need to filter out java-isolated
and java-non-isolated tests. New version of test.py simply reads
list of java tests to be executed from modules/java-tests/test_commands,
generated during build time.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>
Message-Id: <20210614042047.13437-3-jwkozaczuk@gmail.com>