Fix warnings about misleading indentation
Gcc 11 found several places where the code has indentation which may
hint of a bug. For example:

	if (something)
		onething;
		anotherthing;

In such code, the indentation suggests that the programmer indented
for both things to be done only if something is true - and yet the
lack of braces mean just onething is under the condition, and
anotherhing is run unconditionally!

I fixed the indentation or the code, as it seems to me the intention
was. In some cases it seems gcc found real bugs in the code (however,
these bugs seem in very obscure cases).

This issue reminds us why OSv's coding conventions require that "if"
must always use braces - even for a single statement (most of the
involved code predates this coding convention).

Signed-off-by: Nadav Har'El <nyh@scylladb.com>

x

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20210614062057.1998552-5-nyh@scylladb.com>