Makefile: avoid warning in one Musl source file
gcc 11 warns for musl/src/prng/seed48.c that seed48()'s signature in
the header file has a 3-char array, but the implementation has a char
pointer. This is correct, but gcc now warns about such inconsistencies,
and we turn those warnings into errors. Because this is an original
Musl source file that we don't want to needlessly change, let's just
avoid the warning with CFLAGS += -Wno-array-parameter for this specific
file.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20210614062057.1998552-7-nyh@scylladb.com>