fs/vfs/main.cc: fix gcc 11 build error
In fs/vfs/main.cc we have bootfs_start, which was defined as a single
character (extern char bootfs_start) but was really treated as the
beginning of a longer memory area. Gcc 11 can recognize that we access
beyond this single character, and fail the compilation.

To work around this problem, we declare bootfs_start as
	extern char bootfs_start[]

Little needs to be changed besides this - except changing
&bootfs_start + offset - which gcc doesn't like, to
&bootfs_start[offset].

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20210614062057.1998552-8-nyh@scylladb.com>