musl: reinstate ___errno_location as public
The previous patch made ___errno_location hidden to match it with
latest musl. Normally ___errno_location would be an internal symbol
used by setenv and unsetenv for example. However OSv comes with an
libenviron.so that supports namespaces functionality (see
1d3645b6c1a119ea34e1a3dca6ae814d4902905f). The libenviron.so links
number of musl and libc/* objects that reference ___errno_location
through errno macro so we need to keep it exposed as public.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>