prepare to hide kernel symbols: add macros to expose and hide symbols
This patch is the first one in the series that annotates relevant
symbols across number of source files in order to expose or hide them
when kernel is compiled to hide non-glibc symbols. Please note that none
of this patches neither hides nor exposes any symbols. In essense it will
only have an effect once we make changes to the makefile to allow compiling
the kernel with most symbols hidden.

This patch adds new header - export.h - with new macros that either
expose (__visibility__("default")) or hide (__visibility__("hidden"))
the symbols. Please read the comments in the header file for more
details.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>