prepare to hide kernel symbols: expose libc symbols in certain files
This 2nd patch in the series annotates number of standard glibc symbols
across number of source files, mostly under bsd, core, fs and libc
folders. This does not have any effect until we introduce the changes to
the makefile to support compiling kernel with most symbols hidden.

We choose to annotate the symbols that need to be exposed because the relevant files
will be compiled with the flag '-fvisibility=hidden' which would hide
all symbols unless annotated otherwise which takes precedence. This
approach helps to minimize the number of symbols that need to be annotated.
 certain files

Comparing to the version 1 of this patch, this one differs in the following
ways:
- removes changes made to bsd/sys/netinet/in.n, bsd/sys/rpc/xdr.h,
  bsd/sys/sys/md5.h and bsd/sys/sys/libkern.h to avoid exposing symbols
  originating from FreeBSD kernel that may not be completely compatible
  with Linux
- moves some of the OSV_*API annotation from header files to the source
  files

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>