stdio: fix setvbuf() to not set a weird buf size
When we upgraded our Musl version, we got a newer version of setvbuf()
changed in Musl commit 0b80a7b04 in 2018. The old version of setvbuf()
silently ignored the user's buffer and continued to use stdio's default
buffer - a 1024-byte buffer. The new version tried to be better, and
use the user-supply buffer, but created a problem:

Musl's internal implementation uses the first UNGET (=8) bytes of the
buffer for ungetc() support. When Musl wants a 1024-byte buffer, they
actually allocate 8+1024. In order not to change that implementation
detail, their setvbuf() subtracted 8 bytes from the user's supplied
buffer and its length. So if the user provided a 131072 byte (128 KB)
buffer, actually only 131064 bytes were used as the buffer size. This
causes problems for reading from block devices - where the size of
the buffer must be a multiple of BSIZE, and no longer is - as exposed
in issue #1180.

The ugly fix in this patch is to subtract 512, instead of 8, from the
user's given buffer size. A better fix would have been to move the
8-byte unget buffer outside the main buffer, or alternatively for
setvbuf() to allocate the buffer itself (and remember to free it later)
but both fixes would require much more elaborate changes to Musl's
stdio, which I didn't want to do.

The code in this patch is a modified version of
musl_1.1.24/src/stdio/setvbuf.c.

Fixes #1180.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20211129123411.971875-1-nyh@scylladb.com>