aarch64 linker: enforce some sections kept when garbage collection on
This patch changes the aarch64 linker script to keep some sections
when garbage collection (--gc-sections) is on when linking the kernel.

This mostly involves adding KEEP directived to .fixup, .eh_frame*,
.tracepoint_patch_sites* and .init* sections to make them retained
when linking with garbage collection on.

This changes to not have any effect on the linker result when garbage
collection is off - the default.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>