zfs: expose some symbols in solaris part of bsd code
This patch annotates ~90 symbols with explicit public visibility
across various parts of ZFS or related code in the bsd/ subtree. This is
in preparation of the next patch that extracts ZFS code into a separate
libsolaris.so library where all symbols but the ones marked here as public.
These symbols need to be exposed for the main user of libsolaris.so - libzfs.so.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>