tests: do not include tst-solaris-taskq.so as part of ROFS image
With the commit https://github.com/cloudius-systems/osv/commit/d7a627484b34aa8254d8fe444972aa18fe007ca7
the ZFS part of the bsd/ tree is no longer part of the kernel.
The unit test tst-solaris-taskq.so that exercises code that is now
part of the libsolaris.so library should only get included and run
as part of the ZFS image of the tests.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>